+++ R/tostring.R
@@ -38,20 +38,32 @@ default_hsep <- d_hsep_factory()
 #' @rdname tostring
 #'
 #' @inheritParams MatrixPrintForm
-#' @param widths numeric (or NULL). (proposed) widths for the columns of
-#'  \code{x}. The expected length of this numeric vector can be retrieved with
-#'  `ncol() + 1` as the column of row names must also be considered.
-#' @param hsep character(1). Characters to repeat to create header/body
-#'  separator line.
-#' @param tf_wrap logical(1). Should the texts for title, subtitle,
-#'  and footnotes be wrapped?
-#' @param max_width integer(1) (or NULL). If NULL, it will be assigned to
-#'  `options("width")`. Otherwise, it is also possible to use `c("auto")` to
-#'  automatically use the table width.
+#' @param widths numeric (or  NULL). (proposed) widths for the columns
+#'     of \code{x}. The expected length  of this numeric vector can be
+#'     retrieved with  `ncol() + 1`  as the  column of row  names must
+#'     also be considered.
+#' @param  hsep  character(1).   Characters  to  repeat   to  create
+#'     header/body separator line.
+#' @param tf_wrap logical(1). Should  the texts for  title, subtitle,
+#'     and footnotes be wrapped?
+#' @param max_width integer(1), character(1) or NULL. Width that title
+#'     and   footer   (including   footnotes)  materials   should   be
+#'     word-wrapped to. If NULL, it is  set to the current print width
+#'     of the  session (`getOption("width")`). If set to `"auto"`,
+#'     the width of the table (plus any table inset) is used. Ignored
+#'     completely if `tf_wrap` is `FALSE`.
 #'
-#' @details When wrapping texts such as title, subtitle, or footnotes, please
-#'  consider inserting a vector of strings, and every element will be on a
-#'  separate line. Using the newline character is not adviced.
+#' @details  When `tf_wrap` is  `TRUE`, word wrapping happens  as with
+#'     \link[base:strwrap]{base::strwrap}    with     the    following
+#'     exception: individual  words which are longer  than `max_width`
+#'     are  broken  up  in a  way  that  fits  with  the rest  of  the
+#'     wordwrapping.
+#'
+#' Manual insertion of newlines is not supported when `tf_wrap` is on
+#' and will result in a warning and undefined wrapping behavior. Passing
+#' vectors of already split strings remains supported, however in this
+#' case each string is word-wrapped separately with the behavior
+#' described above.
 #'
 #' @examples
 #' mform <- basic_matrix_form(mtcars)
@@ -69,21 +81,22 @@ setMethod("toString", "MatrixPrintForm", function(x,
                                                   hsep = default_hsep()) {
     mat <- x
 
-    ## we create a matrix with the formatted cell contents
-    ##  mat <- matrix_form(x, indent_rownames = TRUE)
-
-    # Text wrapping checks
-    stopifnot(is.logical(tf_wrap))
-    if (tf_wrap) {
-      if (is.null(max_width)) max_width <- options("width")
-      if (!is.character(max_width)) {
-        stopifnot(is.numeric(max_width), max_width > 0)
-      }
-    }
-
+    inset <- table_inset(mat)
     if (is.null(widths)) {
         widths <- propose_column_widths(x)
     }
+
+    ncchar <-  sum(widths) + (length(widths) - 1) * col_gap
+    ## Text wrapping checks
+    stopifnot(is.logical(tf_wrap))
+    if (tf_wrap) {
+        if (is.null(max_width))
+            max_width <- getOption("width", 80L)
+        else if(is.character(max_width) && identical(max_width, "auto"))
+            max_width = ncchar + inset
+        stopifnot(is.numeric(max_width), max_width > 0)
+    }
+
 
     stopifnot(length(widths) == ncol(mat$strings))
 
@@ -95,7 +108,7 @@ setMethod("toString", "MatrixPrintForm", function(x,
     ##    ri <- mat$row_info
     ref_fnotes <- mat$ref_footnotes
 
-    inset <- table_inset(mat)
+
 
     nr <- nrow(body)
     nl_header <- attr(mat, "nlines_header")
@@ -125,9 +138,6 @@ setMethod("toString", "MatrixPrintForm", function(x,
                                         # apply(content, 1, function(x) sum(nchar(x), na.rm = TRUE))
 
     gap_str <- strrep(" ", col_gap)
-
-    ncchar <-  sum(widths) + (length(widths) - 1) * col_gap
-    if (is.character(max_width) && max_width == "auto") max_width <- ncchar
 
     div <- substr(strrep(hsep, ncchar), 1, ncchar)
     txt_head <- apply(head(content, nl_header), 1, .paste_no_na, collapse = gap_str)
@@ -163,24 +173,28 @@ setMethod("toString", "MatrixPrintForm", function(x,
 
 
     allts <- all_titles(x)
-    new_line_warning(allts)
-
-    # Wrapping titles if they go beyond the horizontally allowed space
-    if (tf_wrap) {
-      allts <- unlist(sapply(as.list(allts), wrap_txt, max_width = max_width))
-    }
-
-    titles_txt <- if(any(nzchar(allts))) c(allts, "", .do_inset(div, inset))  else NULL
 
     allfoots <- list("main_footer" = main_footer(x),
                      "prov_footer" = prov_footer(x),
                      "ref_footnotes" = ref_fnotes)
     allfoots <- allfoots[!sapply(allfoots, is.null)]
-    new_line_warning(allfoots)
+
+
+    # Wrapping titles if they go beyond the horizontally allowed space
+    if (tf_wrap) {
+        new_line_warning(allts)
+        allts <- wrap_txt(allts, max_width = max_width)
+    }
+
+    titles_txt <- if(any(nzchar(allts))) c(allts, "", .do_inset(div, inset))  else NULL
 
     # Wrapping footers if they go beyond the horizontally allowed space
     if (tf_wrap) {
-      allfoots <- sapply(allfoots, wrap_txt, max_width = max_width)
+        new_line_warning(allfoots)
+        allfoots$main_footer <- wrap_txt(allfoots$main_footer, max_width - inset)
+        allfoots$ref_footnotes <- wrap_txt(allfoots$ref_footnotes, max_width - inset)
+        ## no - inset here because the prov_footer is not inset
+        allfoots$prov_footer <- wrap_txt(allfoots$prov_footer, max_width)
     }
 
     paste0(paste(c(titles_txt,
@@ -217,7 +231,8 @@ setMethod("toString", "MatrixPrintForm", function(x,
     footer_txt <- .do_inset(rfn, inset)
     if(any(nzchar(footer_txt)))
         footer_txt <- .inset_div(footer_txt, div, inset)
-    if(any(nzchar(footers_v))) {
+    if(any(vapply(footers_v, function(x) any(nzchar(x)),
+                  TRUE))) {
         if(any(nzchar(prvf)))
             provtxt <- c(if(any(nzchar(fter))) "",
                          prvf)
@@ -245,62 +260,110 @@ setMethod("toString", "MatrixPrintForm", function(x,
 
 new_line_warning <- function(str_v) {
   if (any(unlist(sapply(str_v, grepl, pattern = "\n")))) {
-    msg <- c("Newline was manually inserted with its ASCII code. We suggest and ",
-             "support adding newlines by splitting title, subtitle or footnotes ",
-             "in a vector of strings, so that every element will be on a separate line.")
+      msg <- c("Detected manual newlines when automatic title/footer word-wrapping is on.",
+               "This is unsupported and will result in undefined behavior. Please either ",
+               "utilize automatic word-wrapping with newline characters inserted, or ",
+               "turn off automatic wrapping and wordwrap all contents manually by inserting ",
+               "newlines.")
     warning(paste0(msg, collapse = ""))
   }
 }
 
-wrap_txt <- function(txt, max_width) {
-  if (any(sapply(txt, function(x) nchar(x) > max_width))) {
-
-    # Checking if any word is larger than max_width
-    spl_txt_l <- unname(sapply(txt, function(x) {
-      if (nchar(x) > 0) {
-        strsplit(x, split = " ")
-      } else {
-        x
-      }
-    }))
-    which_large_txt <- sapply(spl_txt_l, function(x) any(nchar(x) > max_width))
-
-    if (any(which_large_txt)) {
-      warning(
-        "Found the following txt that could not be wrapped with max_width of ",
-        max_width,
-        ":\n- ",
-        paste0(txt[which_large_txt], collapse = "\n- "),
-        "\nAttempting at a split after ", max_width, " characters."
-      )
-
-      spl_txt_l[which_large_txt] <- lapply(
-        spl_txt_l[which_large_txt],
-        function(x) {
-          unname(unlist(lapply(x, function(y) {
-            sapply(
-              seq(from = 1, to = nchar(y), by = max_width),
-              function(i) substr(y, i, i + max_width - 1)
-            )
-          })))
+wrap_string <- function(str, max_width) {
+    stopifnot(is.character(str) && length(str) == 1)
+    naive <- strwrap(str, max_width + 1)
+    while(any(nchar(naive) > max_width)) {
+        good <- character()
+        bwi <- which(nchar(naive) > max_width)[1]
+        curbw <- naive[bwi]
+        if (bwi > 2) {
+            good <- c(good, naive[1:(bwi - 2)])
         }
-      )
-      # Paste everything back together
-      txt <- lapply(spl_txt_l, paste, collapse = " ")
+        if(bwi > 1) {
+            str_before <- naive[bwi - 1]
+        } else {
+            str_before <- ""
+        }
+        room <- max_width - nchar(str_before) - (bwi > 1)
+        if(room <= 0) {
+            toadd <- c(str_before, substr(curbw, 1, max_width))
+            room <- 0
+            leftover <- substr(curbw, max_width + 1, nchar(curbw))
+        } else {
+            goodpart <- substr(curbw, 1, room)
+            if(nzchar(str_before))
+                toadd <- paste(str_before, goodpart)
+            else
+                toadd <- goodpart
+            leftover <- substr(curbw, room + 1, nchar(curbw))
+        }
+        good <- c(good, toadd)
+        if(bwi == length(naive))
+            good <- c(good, leftover)
+        else
+            good <- c(good,
+                      paste(leftover, naive[bwi + 1]),
+                      if(bwi < length(naive) - 1) naive[seq(bwi + 2, length(naive))]
+                      )
+        str <- paste(good, collapse = " ")
+        naive <- strwrap(str, max_width + 1)
     }
-    # Main wrapper
-    txt_out <- lapply(txt,
-      strwrap,
-      width = max_width + 1
-    )
-    names(txt_out) <- NULL
-    txt_out <- unlist(txt_out)
-
-    txt_out
-  } else {
-    unlist(txt)
-  }
+    naive
 }
+
+
+wrap_txt <- function(txt, max_width) {
+    unlist(lapply(txt, wrap_string, max_width = max_width), use.names = FALSE)
+}
+
+##   if (any(sapply(txt, function(x) nchar(x) > max_width))) {
+
+##     # Checking if any word is larger than max_width
+##     spl_txt_l <- unname(sapply(txt, function(x) {
+##       if (nchar(x) > 0) {
+##         strsplit(x, split = " ")
+##       } else {
+##         x
+##       }
+##     }))
+##     which_large_txt <- sapply(spl_txt_l, function(x) any(nchar(x) > max_width))
+
+##     if (any(which_large_txt)) {
+##       warning(
+##         "Found the following txt that could not be wrapped with max_width of ",
+##         max_width,
+##         ":\n- ",
+##         paste0(txt[which_large_txt], collapse = "\n- "),
+##         "\nAttempting at a split after ", max_width, " characters."
+##       )
+
+##       spl_txt_l[which_large_txt] <- lapply(
+##         spl_txt_l[which_large_txt],
+##         function(x) {
+##           unname(unlist(lapply(x, function(y) {
+##             sapply(
+##               seq(from = 1, to = nchar(y), by = max_width),
+##               function(i) substr(y, i, i + max_width - 1)
+##             )
+##           })))
+##         }
+##       )
+##       # Paste everything back together
+##       txt <- lapply(spl_txt_l, paste, collapse = " ")
+##     }
+##     # Main wrapper
+##     txt_out <- lapply(txt,
+##       strwrap,
+##       width = max_width + 1
+##     )
+##     names(txt_out) <- NULL
+##     txt_out <- unlist(txt_out)
+
+##     txt_out
+##   } else {
+##     unlist(txt)
+##   }
+## }
 
 pad_vert_top <- function(x, len) {
     c(x, rep("", len - length(x)))