+++ NEWS.md
@@ -1,9 +1,15 @@
+## formatters 0.3.3.7 
+ * `tf_wrap` functionality now breaks up words that are too large with respect to surrounding wordwrapping.
+ * `tf_wrap` behavior no longer emits a warning when breaking up 'words' larger than `max_width`
+ * `tf_wrap` `max_width = "auto"` now sets the `max_width` to `sum(widths) + table_inset(x)` instead of `sum(widths)`
+
 ## formatters 0.3.3.6
- * `toString` now respects `table_inset` values set on a `MatrixPrintForm` object
+ * Experimental `tf_wrap` functionality added to `toString` `MatrixPrintForm` method
 
 ## formatters 0.3.3.5
  * `MatrixPrintForm` class now includes a `table_inset` element
  * Add `table_inset` and `table_inset<-` generics now exported
+ * `toString` now respects `table_inset` values set on a `MatrixPrintForm` object
 
 ## formatters 0.3.3.4
  * Use `exact=TRUE` for all calls to `attr` in accessor functions.